+++ src/lib.rs
@@ -7,28 +7,31 @@
 use std::io::{Write, Read};
 use std::io::Result as IOResult;
 
+pub mod padding;
+pub use padding::{Padding, NoPadding, LengthPadding};
+
 /// **BitWriter** is a writer for single bit values
 ///
 /// Bits will be grouped to a single byte before writing to the inner writer.
 /// The first Bit will be the most significant bit of the byte.
 ///
-/// When dropping this writer, any remaining bits will be written as well as an additional byte
-/// containing how many bits are significant in the last data byte. This means you can write any
-/// number of bits, they do not need to align to multiples of 8.
+/// When dropping this writer, any remaining bits will be written according to the padding used.
+/// The default padding is [NoPadding](struct.NoPadding.html)
 ///
 /// # Examples
 ///
 /// ```
 /// extern crate bitstream;
 ///
-/// let mut vec = Vec::new();
+/// let vec = Vec::new();
 /// let mut bit_writer = bitstream::BitWriter::new(vec);
 ///
 /// assert!(bit_writer.write_bit(true).is_ok());
 /// assert!(bit_writer.write_bit(false).is_ok());
 /// ```
-pub struct BitWriter<W> where W: Write {
+pub struct BitWriter<W, P> where W: Write, P: Padding {
     inner: W,
+    padder: P,
     last_byte: u8,
     last_fill: u8,
 }
@@ -44,7 +47,7 @@ pub struct BitWriter<W> where W: Write {
 /// extern crate bitstream;
 /// use std::io::Cursor;
 ///
-/// let mut vec = vec![192, 2];
+/// let vec = vec![192, 2];
 /// let mut bit_reader = bitstream::BitReader::new(Cursor::new(vec));
 /// let first_read = bit_reader.read_bit();
 /// assert!(first_read.is_ok());
@@ -52,20 +55,30 @@ pub struct BitWriter<W> where W: Write {
 /// assert!(option.is_some());
 /// assert!(option.unwrap());
 /// ```
-pub struct BitReader<R> where R: Read {
+pub struct BitReader<R, P> where R: Read, P: Padding {
+    padder: P,
     inner: R,
     ended: bool,
-    fill: u8,
+    fill: usize,
     current: u8,
-    buffer: [u8; 3],
-    byte_fill: u8,
+    buffer: Box<[u8]>,
+    bits_left: usize,
 }
 
-impl<W> BitWriter<W> where W: Write {
-    /// Create a new BitWriter, writing to the inner writer.
+
+impl<W> BitWriter<W, NoPadding> where W: Write {
+    /// Create a new BitWriter with no padding, writing to the inner writer.
     pub fn new(write: W) -> Self {
+        BitWriter::with_padding(write, NoPadding::new())
+    }
+}
+
+impl<W, P> BitWriter<W, P> where W: Write, P: Padding {
+    /// Create a new BitWriter with the given padding
+    pub fn with_padding(write: W, padder: P) -> Self {
         BitWriter {
             inner: write,
+            padder: padder,
             last_byte: 0,
             last_fill: 0,
         }
@@ -91,39 +104,72 @@ impl<W> BitWriter<W> where W: Write {
     }
 }
 
-impl<W> Drop for BitWriter<W> where W: Write {
+impl<W, P> Drop for BitWriter<W, P> where W: Write, P: Padding {
     fn drop(&mut self) {
-        if self.last_fill > 0 {
-            let _ = self.inner.write_all(&[self.last_byte, self.last_fill]);
-        } else {
-            let _ = self.inner.write_all(&[8u8]);
-        }
+        let _ = self.padder.pad(self.last_byte, self.last_fill, &mut self.inner);
     }
 }
 
 
-impl<R> BitReader<R> where R: Read {
-    /// Create a new BitReader, reading from the inner reader.
+impl<R> BitReader<R, NoPadding> where R: Read {
+    /// Create a new BitReader, with no padding, reading from the inner reader.
     pub fn new(reader: R) -> Self {
+        BitReader::with_padding(reader, NoPadding::new())
+    }
+}
+
+impl<R, P> BitReader<R, P> where R: Read, P: Padding {
+
+    /// Create a new BitReader, using the supplied padding.
+    ///
+    /// This can be used to supply a custom padding to the bit reader.
+    ///
+    /// # Examples
+    /// ```
+    /// extern crate bitstream;
+    /// use std::io::Cursor;
+    ///
+    /// let vec = vec![192, 2];
+    /// let mut bit_reader = bitstream::BitReader::with_padding(Cursor::new(vec),
+    ///                                                         bitstream::LengthPadding::new());
+    /// let _ = bit_reader.read_bit();
+    /// let _ = bit_reader.read_bit();
+    /// let last = bit_reader.read_bit();
+    /// assert!(last.is_ok());
+    /// /// None indicates there is nothing left to read
+    /// assert!(last.unwrap().is_none());
+    /// ```
+    pub fn with_padding(reader: R, padder: P) -> Self {
+        let buf_size = padder.max_size() + 1;
+        let buffer = vec![0; buf_size];
+
         BitReader {
             inner: reader,
+            padder: padder,
             fill: 0,
             ended: false,
-            buffer: [0, 0, 0],
+            buffer: buffer.into_boxed_slice(),
             current: 0,
-            byte_fill: 8,
+            bits_left: 0,
         }
     }
 
     fn fill_buffer(&mut self) -> IOResult<()> {
-        while !self.ended && self.fill != 3 {
-            match self.inner.read(&mut self.buffer[self.fill as usize..]) {
+        while !self.ended && self.fill != self.buffer.len() {
+            match self.inner.read(&mut self.buffer[self.fill..]) {
                 Ok(0) => {
                     self.ended = true;
-                    self.fill -= 1;
-                    self.byte_fill = self.buffer[self.fill as usize];
-                },
-                Ok(n) => self.fill += n as u8,
+                    let buf_pad_start = if self.fill < self.buffer.len() {
+                        0
+                    } else {
+                        1
+                    };
+                    self.bits_left = self.padder.bits_left(&self.buffer[buf_pad_start..self.fill])?;
+                }
+                Ok(n) => {
+                    self.fill += n;
+                    self.bits_left = 8;
+                }
                 Err(e) => return Err(e),
             }
         }
@@ -137,23 +183,27 @@ impl<R> BitReader<R> where R: Read {
     /// # Failures
     /// Will return an error if the inner reader returns one
     pub fn read_bit(&mut self) -> IOResult<Option<bool>> {
-        if self.fill > 0 && self.current == self.byte_fill {
-            self.buffer = [self.buffer[1], self.buffer[2], 0];
-            self.current = 0;
-            self.fill -= 1;
-        }
         self.fill_buffer()?;
-        if self.fill > 0 {
+        if self.bits_left == 0 {
+            Ok(None)
+        } else {
             let res = (self.buffer[0] & (128u8 >> self.current)) == (128u8 >> self.current);
             self.current += 1;
+            self.bits_left -= 1;
+
+            if self.current == 8 {
+                self.current = 0;
+                self.fill -= 1;
+                unsafe {
+                    std::ptr::copy(self.buffer[1..].as_ptr(), self.buffer[..].as_mut_ptr(), self.buffer.len() - 1);
+                }
+            }
             Ok(Some(res))
-        } else {
-            Ok(None)
         }
     }
 }
 
-impl<R> Iterator for BitReader<R> where R: Read {
+impl<R, P> Iterator for BitReader<R, P> where R: Read, P: Padding {
     type Item = bool;
 
     fn next(&mut self) -> Option<Self::Item> {
@@ -170,10 +220,70 @@ mod tests {
     use std::io::Cursor;
 
     #[test]
-    fn test_writer() {
+    fn test_writer_no_pad() {
         let mut vec = Vec::new();
         {
             let mut bit_writer = BitWriter::new(&mut vec);
+            assert!(bit_writer.write_bit(true).is_ok());
+            assert!(bit_writer.write_bit(true).is_ok());
+            assert!(bit_writer.write_bit(false).is_ok());
+            assert!(bit_writer.write_bit(true).is_ok());
+            assert!(bit_writer.write_bit(true).is_ok());
+            assert!(bit_writer.write_bit(false).is_ok());
+            assert!(bit_writer.write_bit(false).is_ok());
+            assert!(bit_writer.write_bit(true).is_ok());
+            assert!(bit_writer.write_bit(true).is_ok());
+            assert!(bit_writer.write_bit(true).is_ok());
+        }
+        assert_eq!(vec.len(), 2);
+        assert_eq!(vec[0], 217);
+        assert_eq!(vec[1], 192);
+    }
+
+    #[test]
+    fn test_writer_no_pad_empty() {
+        let mut vec = Vec::new();
+        {
+            BitWriter::new(&mut vec);
+        }
+        assert_eq!(vec.len(), 0);
+    }
+
+    #[test]
+    fn test_reader_no_pad() {
+        let mut vec = Cursor::new(vec![200, 192]);
+        let mut bit_reader = BitReader::new(&mut vec);
+        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), true);
+        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), true);
+        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), false);
+        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), false);
+        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), true);
+        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), false);
+        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), false);
+        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), false);
+        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), true);
+        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), true);
+        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), false);
+        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), false);
+        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), false);
+        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), false);
+        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), false);
+        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), false);
+        assert!(bit_reader.read_bit().unwrap().is_none());
+    }
+
+    #[test]
+    fn test_reader_no_pad_empty() {
+        let mut vec = Cursor::new(&[]);
+        let mut bit_reader = BitReader::new(&mut vec);
+        assert!(bit_reader.read_bit().unwrap().is_none());
+    }
+
+    #[test]
+    fn test_writer_length_pad() {
+        let mut vec = Vec::new();
+        {
+            let mut bit_writer = BitWriter::with_padding(&mut vec, LengthPadding::new());
             assert!(bit_writer.write_bit(true).is_ok());
             assert!(bit_writer.write_bit(true).is_ok());
             assert!(bit_writer.write_bit(false).is_ok());
@@ -192,27 +302,33 @@ mod tests {
     }
 
     #[test]
-    fn test_reader() {
-        let mut vec = Cursor::new(vec![200, 192, 2]);
-        let mut bit_reader = BitReader::new(&mut vec);
-        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), true);
-        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), true);
-        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), false);
-        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), false);
-        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), true);
-        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), false);
-        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), false);
-        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), false);
-        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), true);
-        assert_eq!(bit_reader.read_bit().unwrap().unwrap(), true);
-        assert!(bit_reader.read_bit().unwrap().is_none());
+    fn test_writer_length_pad_empty() {
+        let mut vec = Vec::new();
+        {
+            BitWriter::with_padding(&mut vec, LengthPadding::new());
+        }
+        assert_eq!(vec.len(), 1);
+        assert_eq!(vec[0], 8);
     }
 
     #[test]
-    fn test_write_read() {
+    fn test_write_read_length_pad_empty() {
         let mut vec = Vec::new();
         {
-            let mut bit_writer = BitWriter::new(&mut vec);
+            BitWriter::with_padding(&mut vec, LengthPadding::new());
+        }
+        {
+            let mut cur = Cursor::new(&vec);
+            let mut bit_reader = BitReader::with_padding(&mut cur, LengthPadding::new());
+            assert!(bit_reader.read_bit().unwrap().is_none());
+        }
+    }
+
+    #[test]
+    fn test_write_read_length_pad() {
+        let mut vec = Vec::new();
+        {
+            let mut bit_writer = BitWriter::with_padding(&mut vec, LengthPadding::new());
             assert!(bit_writer.write_bit(true).is_ok());
             assert!(bit_writer.write_bit(true).is_ok());
             assert!(bit_writer.write_bit(false).is_ok());
@@ -226,7 +342,7 @@ mod tests {
         }
         {
             let mut cur = Cursor::new(&vec);
-            let mut bit_reader = BitReader::new(&mut cur);
+            let mut bit_reader = BitReader::with_padding(&mut cur, LengthPadding::new());
             assert_eq!(bit_reader.read_bit().unwrap().unwrap(), true);
             assert_eq!(bit_reader.read_bit().unwrap().unwrap(), true);
             assert_eq!(bit_reader.read_bit().unwrap().unwrap(), false);