+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Change Log
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/) 
+The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [0.2.0] -- 2017.06.04
+### Added
+- Custom padding options via `Padding` trait
+- Implemented `NoPadding` and `LengthPadding` as out-of-the-box padding options
+### Changed
+- `BitReader` and `BitWriter` signature changed
+- `NoPadding` is the new default padding
 
 ## [0.1.2] -- 2017.01.06
 ### Added